requirements.txt
streamlit
ultralytics
opencv-python-headless==4.10.0.84
Pillow
numpy

